refactor(padi): L7 rename — @kolu/terminal-workspace → @kolu/terminal-vocab
Its own commit (separate from the motion) so it can be reverted to keep-name
with one revert if preferred — the motion in the parent commit stands either way.

After L7's motion the package holds only the browser-safe TerminalSnapshot
schema family + agentProjection — the shared terminal vocabulary the client
reads. "workspace" named the old assembly (sensors · fold · fs/git endpoint),
which now lives in padi; the name should say what the leaf now is. "vocab"
matches the in-repo precedent (padi/src/vocab.ts).

Pure identity rename, zero logic changes: package dir + name, every import
specifier, the 5 workspace deps, the buildId.closure HASHED_ROOTS key/dir, and
default.nix's padiSrc path. Conceptual/domain prose ("terminal-workspace
daemon/surface/contract") and the Atlas notes are left untouched — that's L8's
doc-sweep and the Atlas is historical record.

Gates green on the renamed tree: whole-workspace typecheck + biome, full unit
suites (terminal-vocab 1 · padi 34 · common 4 · padi-tui 3 · client 69 ·
server 18+1skip), seal 6/6, padi.drv evaluates.

Ledger: docs/atlas padi-cleanup.mdx §L7.
