+++ default.nix
@@ -94,7 +94,7 @@ let
       ./packages/terminal-themes
       ./packages/theme
       ./packages/memorable-names
-      ./packages/terminal-workspace
+      ./packages/terminal-vocab
       ./packages/terminal-protocol
       ./packages/kaval
       ./packages/kaval-tui
@@ -266,11 +266,11 @@ let
       (padiPkgRoot ./packages/terminal-protocol)
       (padiPkgRoot ./packages/surface-daemon)
       (padiPkgRoot ./packages/surface-daemon-supervisor)
-      # terminal-workspace — the browser-safe TerminalSnapshot vocabulary +
+      # terminal-vocab — the browser-safe TerminalSnapshot vocabulary +
       # agentProjection padi's closure reaches (L7 folded the node-only sensors +
       # fold + fs/git endpoint INTO padi/src/terminalWorkspace, so what remains
       # here is the shared leaf; the whole src still hashes into padi's key).
-      (padiPkgRoot ./packages/terminal-workspace)
+      (padiPkgRoot ./packages/terminal-vocab)
       # The domain leaves padi's closure reaches: serving, the agent/forge/git
       # integrations, transcripts, and the shared utilities. (`@kolu/surface` and
       # the npm deps are NOT here — surface is the framework "electricity" (a
