+++ packages/common/package.json
@@ -31,7 +31,7 @@
     "@kolu/surface-app": "workspace:*",
     "@kolu/surface-daemon-supervisor": "workspace:*",
     "@kolu/surface-map": "workspace:*",
-    "@kolu/terminal-workspace": "workspace:*",
+    "@kolu/terminal-vocab": "workspace:*",
     "@orpc/contract": "^1.13.13",
     "anyagent": "workspace:*",
     "anyforge": "workspace:*",
