+++ packages/integrations/github/src/index.ts
@@ -3,7 +3,7 @@
  *  Spawns `gh pr view` (via `KOLU_GH_BIN`), classifies its failures, and
  *  derives the check rollup. The forge-neutral wire schemas, poll loop, and
  *  remote-host grammar live in `anyforge`; host-to-forge detection lives in
- *  the sensor layer (`@kolu/terminal-workspace`). This package never sees its
+ *  the sensor layer (`@kolu/terminal-vocab`). This package never sees its
  *  sibling adapters. */
 
 export { classifyGhError, deriveCheckStatus, extractChecks } from "./github.ts";
