+++ packages/padi/package.json
@@ -31,7 +31,7 @@
     "@kolu/surface": "workspace:*",
     "@kolu/surface-daemon": "workspace:*",
     "@kolu/surface-daemon-supervisor": "workspace:*",
-    "@kolu/terminal-workspace": "workspace:*",
+    "@kolu/terminal-vocab": "workspace:*",
     "@orpc/client": "^1.13.13",
     "@orpc/contract": "^1.13.13",
     "@orpc/experimental-publisher": "^1.13.13",
