+++ packages/server/package.json
@@ -23,7 +23,7 @@
     "@kolu/surface-daemon-supervisor": "workspace:*",
     "@kolu/surface-remote": "workspace:*",
     "@kolu/terminal-protocol": "workspace:*",
-    "@kolu/terminal-workspace": "workspace:*",
+    "@kolu/terminal-vocab": "workspace:*",
     "@orpc/client": "^1.13.13",
     "@orpc/contract": "^1.13.13",
     "@orpc/experimental-pino": "^1.13.13",
