+++ packages/server/src/remotePadiSsh.test.ts
@@ -46,7 +46,7 @@ import {
 import { isContractVersionCompatible } from "@kolu/surface/define";
 import { firstFrameOrUndefined } from "@kolu/surface/first-frame";
 import { isLocalHost, makeSession, sshConnector } from "@kolu/surface-remote";
-import type { TerminalId } from "@kolu/terminal-workspace/schema";
+import type { TerminalId } from "@kolu/terminal-vocab/schema";
 import { afterAll, describe, expect, it } from "vitest";
 
 const SSH_HOST = process.env.KOLU_E2E_SSH_HOST;
