+++ packages/solid-statepip/package.json
@@ -24,7 +24,7 @@
     "vitest": "^4.1.0"
   },
   "dependencies": {
-    "@kolu/terminal-workspace": "workspace:*",
+    "@kolu/terminal-vocab": "workspace:*",
     "solid-js": "^1.9.0"
   }
 }
