+++ packages/terminal-vocab/README.md
