+++ packages/terminal-vocab/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@kolu/terminal-workspace",
+  "name": "@kolu/terminal-vocab",
   "version": "0.1.0",
   "private": true,
   "license": "AGPL-3.0-or-later",
