+++ packages/terminal-vocab/src/agentProjection.test.ts
