+++ packages/terminal-vocab/src/agentProjection.ts
