+++ packages/terminal-vocab/src/schema.ts
