+++ packages/terminal-vocab/tsconfig.json
